break up the libcontainer driver into two parts
Make it explicit that one part of the driver (--driver libcontainer) executes
outside of the container (it exists mostly just to call namespaces.Exec), and
the other part (--driver libcontainer-init) executes inside the container,
acting as PID=1 (init), and it is responsible for calling namespaces.Init().

The libcontainer-init driver re-execs as the abspath driver right after doing
the necessary setup inside the container.

Container configuration is sent to libcontainer-driver as a gob-serialized
struct through a unix pipe.